[ISB-362] Added consumer + service method to process Wallet and send to Timeline + added mapper for different QueueOperationDTO

+++ src/main/java/it/gov/pagopa/wallet/model/Wallet.java
@@ -25,7 +25,7 @@ public class Wallet {
   private BigDecimal amount;
   private BigDecimal accrued;
   private BigDecimal refunded;
-  private int nTrx;
+  private Long nTrx;
   private int nInstr;
   private LocalDateTime unsubscribeDate;
 }
